requirements.txt
absl-py==1.1.0
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
alpaca-trade-api==2.1.0
altair==4.2.0
anyio==3.5.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
Authlib==0.15.5
Babel==2.10.1
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bidict==0.22.0
bleach==5.0.0
blinker==1.4
boto3==1.20.54
botocore==1.23.54
Brotli==1.0.9
bs4==0.0.1
cachetools==5.0.0
ccxt==1.81.95
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
commonmark==0.9.1
cryptography==3.4.8
cycler==0.11.0
dateparser==1.1.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dnspython==2.2.1
email-validator==1.2.1
entrypoints==0.4
eventlet==0.33.0
executing==0.8.3
extra-streamlit-components==0.1.53
fastjsonschema==2.15.3
Flask==2.1.2
Flask-Babel==2.0.0
Flask-Compress==1.12
Flask-Gravatar==0.5.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.4
Flask-SocketIO==5.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==1.12
fonttools==4.33.3
frozenlist==1.3.0
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.46.3
h5py==3.7.0
httpagentparser==1.9.2
hydralit-components==1.0.9
idna==2.10
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
json5==0.9.8
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.13.0
jupyterlab-widgets==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
ldap3==2.9.1
libclang==14.0.1
lxml==4.8.0
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.3
multidict==6.0.2
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numerize==0.12
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pandas-ta==0.3.14b0
pandocfilters==1.5.0
paramiko==2.10.4
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pgadmin4==6.8
pickleshare==0.7.5
Pillow==9.1.0
plotly==5.8.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.1.2
pycparser==2.21
pydeck==0.7.1
Pygments==2.12.0
PyJWT==2.3.0
Pympler==1.0.1
PyNaCl==1.5.0
pyotp==2.6.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-binance==1.0.16
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
qrcode==7.3.1
regex==2022.3.2
requests==2.25.1
requests-oauthlib==1.3.1
rich==12.4.4
rsa==4.8
s3transfer==0.5.2
scikit-learn==1.1.1
scipy==1.8.1
semver==2.13.0
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
speaklater3==1.4
SQLAlchemy==1.4.36
sqlparse==0.4.2
sshtunnel==0.4.0
stack-data==0.2.0
streamlit==1.10.0
streamlit-authenticator==0.1.4
streamlit-option-menu==0.3.2
tenacity==8.0.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.13.3
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
twilio==7.9.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
ua-parser==0.10.0
ujson==5.2.0
urllib3==1.26.9
user-agents==2.2.0
validators==0.19.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
websockets==10.3
Werkzeug==2.0.3
widgetsnbextension==3.6.0
wrapt==1.14.1
WTForms==3.0.1
yarl==1.7.2
zipp==3.8.0
